+++ slider/data/dashboard-template.jsonnet
@@ -13,7 +13,7 @@ local sloTarget =
     valueName='name',
   ).addTarget(
     prometheus.target(
-      'slo:info:target',
+      std.extVar("SLO_INFO_TARGET") + '{id="$service_id"}',
       legendFormat='{{ target }}',
     )
   );
@@ -26,7 +26,7 @@ local sloWindow =
     valueName='name',
   ).addTarget(
     prometheus.target(
-      'slo:info:target',
+      std.extVar("SLO_INFO_TARGET") + '{id="$service_id"}',
       legendFormat='{{ window }}',
     )
   );
@@ -45,13 +45,31 @@ local sloSuccessRatioGraph =
     },
   ).addTarget(
     prometheus.target(
-      'slo:info:target',
+      std.extVar("SLO_INFO_TARGET") + '{id="$service_id"}',
       legendFormat='Target',
     )
   ).addTarget(
     prometheus.target(
-      'slo:success:ratio',
+      std.extVar("SLO_SUCCESS_RATIO") + '{id="$service_id"}',
       legendFormat='Current',
+    )
+  );
+
+local errorBudgetLeftGraph = 
+  graphPanel.new(
+    title='Error Budget Left',
+    datasource='Prometheus',
+    linewidth=2,
+    format='percentunit',
+    decimals=2,
+    fill=100,
+    aliasColors={
+      'Error Budget left over SLO period': 'light-green',
+    },
+  ).addTarget(
+    prometheus.target(
+      '((1 - scalar(' + std.extVar("SLO_INFO_TARGET") + '{id="$service_id"})' + ') - (sum(sum_over_time(' + std.extVar("SLO_TOTAL_QUERY") + '['+std.extVar("SLO_WINDOW") +'])) - sum(sum_over_time(' + std.extVar("SLO_GOOD_QUERY") + '['+std.extVar("SLO_WINDOW") +'])) or vector(0)) / sum(sum_over_time(' + std.extVar("SLO_TOTAL_QUERY") + '['+std.extVar("SLO_WINDOW") +']))) / (1- scalar(' + std.extVar("SLO_INFO_TARGET") + '{id="$service_id"}))',
+      legendFormat='Error Budget left over SLO period',
     )
   );
 
@@ -80,6 +98,14 @@ dashboard.new(
     refresh='time',
   )
 )
+.addTemplate(
+  template.new(
+    'service_id',
+    '$PROMETHEUS_DS',
+    'label_values(id)',
+    label='Service ID',
+  )
+)
 
 .addPanels(
   [
@@ -88,5 +114,7 @@ dashboard.new(
     sloWindow { gridPos: { h: 4, w: 4, x: 3, y: 0 } },
 
     sloSuccessRatioGraph { gridPos: { h: 8, w: 7, x: 0, y: 4 } },
+
+    errorBudgetLeftGraph { gridPos: { h: 12, w: 14, x: 7, y: 0 } },
   ]
 )
